requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
psycopg2-binary==2.9.7
python-telegram-bot==20.6
cloudinary==1.36.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
